backend/requirements.txt
astroid==2.4.2
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
et-xmlfile==1.0.1
fastapi==0.61.2
future==0.18.2
h11==0.11.0
imageio==2.9.0
imutils==0.5.3
isort==5.5.4
jdcal==1.4.1
joblib==0.17.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.5
numpy==1.19.2
odfpy==1.4.1
opencv-python==4.4.0.44
openpyxl==3.0.5
pandas==1.1.3
pathlib==1.0.1
Pillow==7.2.0
pydantic==1.7.2
pylint==2.6.0
pyparsing==2.4.7
pytesseract==0.3.6
python-bidi==0.4.2
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
starlette==0.13.6
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.1
uvicorn==0.12.2
wrapt==1.12.1
xlrd==1.2.0